Clean up
Remove document_generator.py and markdown_generator.py since they're no longer
being used.

Rename x -> templates.
